requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-on-heroku==1.1.2
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pypng==0.0.21
PyQRCode==1.2.1
python-magic==0.4.27
python-magic-bin==0.4.14
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
